run with explicit config and leave simulation files if flagged

+++ htsohm/simulation/gas_loading.py
@@ -7,7 +7,6 @@ from datetime import datetime
 from string import Template
 from pathlib import Path
 
-from htsohm import config
 from htsohm.material_files import write_mol_file, write_mixing_rules
 from htsohm.material_files import write_pseudo_atoms, write_force_field
 from htsohm.simulation.files import load_and_subs_template
@@ -106,7 +105,7 @@ def parse_output(output_file, material, simulation_config):
                 gas_loading.host_adsorbate_vdw  = float(line.split()[5])
                 gas_loading.host_adsorbate_cou  = float(line.split()[7])
             line_counter += 1
-        
+
     material.gas_loading.append(gas_loading)
 
 def pressure_string(p):
@@ -115,7 +114,7 @@ def pressure_string(p):
     else:
         return str(p)
 
-def run(material, simulation_config):
+def run(material, simulation_config, config):
     """Runs gas loading simulation.
 
     Args:
@@ -171,12 +170,13 @@ def run(material, simulation_config):
             while not Path(output_path).exists():
                 subprocess.run(["simulate", "-i", "./{}_loading.input".format(adsorbate)],
                     check = True, cwd = output_dir)
-        
+
             print("Output directory : {}".format(output_dir))
 
             # Parse output
             parse_output(output_path, material, simulation_config)
-            shutil.rmtree(output_dir, ignore_errors=True)
+            if not config['keep_configs']:
+                shutil.rmtree(output_dir, ignore_errors=True)
             sys.stdout.flush()
         except FileNotFoundError as err:
             print(err)
